# Added Develocity configuration and enabled remote build cache.