@ribafish 2025-03-12
Replaced by a real PR: https://github.com/eclipse-edc/Connector/pull/4862